error arreglado, no funcionan correctamente los metodos

+++ src/main/java/com/AloBomnito/AdminRepository/copy/AdministradorRepository.java
@@ -5,7 +5,7 @@ import com.AloBomnito.modelo.copy.Administrador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface AdministradorRepository extends JpaRepository<Administrador, Integer>{
+public interface AdministradorRepository extends JpaRepository<Administrador, Long>{
 	
 
 }
